Fazendo carregamento nas requisições e personalizando mensagem de não autorização

+++ front-end/src/empresa/ListaEmpresas.js
@@ -3,6 +3,7 @@ import {FlatList,Alert,TextInput,SafeAreaView,StyleSheet} from "react-native";
 import {useIsFocused} from "@react-navigation/native";
 import {Menu,Provider} from "react-native-paper";
 import {ListItem,Button} from "@rneui/themed";
+import Espera from "../Espera.js";
 import {estadoInicialMenu} from "../store/config.js";
 import {reducer} from "../store/menuReducer.js";
 import {menuAtivo,menuInativo} from "../store/menuAction.js";
@@ -13,6 +14,7 @@ export default function ListaEmpresas(props) {
     const [empresas,setEmpresas] = useState([]);
     const [pesquisa,setPesquisa] = useState("");
     const [stateMenu,dispatchMenu] = useReducer(reducer,estadoInicialMenu);
+    const [esperar,setEsperar] = useState(false);
     const voltouFoco = useIsFocused();
 
     useEffect(() => {
@@ -32,40 +34,54 @@ export default function ListaEmpresas(props) {
 
     async function obtemEmpresas() {
         try {
+            setEsperar(true);
             const token = await AsyncStorage.getItem("token");
-            const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
-            const opcoes = {method: "GET",headers: cabecalho};
-            const resposta = await fetch(urlEmpresa,opcoes);
-            let msg = await obtemMensagemErro(resposta);
 
-            if (msg && msg !== "")
-                throw new Error(msg);
+            if (token && token !== "") {
+                const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
+                const opcoes = {method: "GET",headers: cabecalho};
+                const resposta = await fetch(urlEmpresa,opcoes);
+                let msg = await obtemMensagemErro(resposta);
 
-            const dados = await resposta.json();
-            setEmpresas(dados);
+                if (msg && msg !== "")
+                    throw new Error(msg);
+
+                const dados = await resposta.json();
+                setEmpresas(dados);
+            }
         }
         catch(erro) {
             Alert.alert("Empresa",JSON.parse(erro.message).mensagem);
+        }
+        finally {
+            setEsperar(false);
         }
     }
 
     async function filtrar() {
         try {
+            setEsperar(true);
             const token = await AsyncStorage.getItem("token");
-            const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
-            const opcoes = {method: "GET",headers: cabecalho};
-            const resposta = await fetch(urlEmpresa + "?nome=" + pesquisa,opcoes);
-            
-            let msg = await obtemMensagemErro(resposta);
-                
-            if (msg && msg !== "")
-                throw new Error(msg);
 
-            const dados = await resposta.json();
-            setEmpresas(dados);
+            if (token && token !== "") {
+                const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
+                const opcoes = {method: "GET",headers: cabecalho};
+                const resposta = await fetch(urlEmpresa + "?nome=" + pesquisa,opcoes);
+                
+                let msg = await obtemMensagemErro(resposta,props.navigation);
+                    
+                if (msg && msg !== "")
+                    throw new Error(msg);
+
+                const dados = await resposta.json();
+                setEmpresas(dados);
+            }
         }
         catch (erro) {
             Alert.alert("Empresa",JSON.parse(erro.message).mensagem);
+        }
+        finally {
+            setEsperar(false);
         }
     }
 
@@ -75,22 +91,29 @@ export default function ListaEmpresas(props) {
                 text: "Sim",
                 async onPress() {
                     try {
+                        setEsperar(true);
                         const token = await AsyncStorage.getItem("token");
-                        const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
-                        let opcoes =  {method: "DELETE",body: empresa,headers: cabecalho};
-                        let resposta = await fetch(urlEmpresa + "/" + empresa.id,opcoes);
-                        let msg = await obtemMensagemErro(resposta);
-                
-                        if (msg && msg !== "")
-                            throw new Error(msg);
 
-                        await obtemEmpresas();
-                        
-                        msg = "Empresa excluída com sucesso.";
-                        Alert.alert("Empresa",msg);
+                        if (token && token !== "") {
+                            const cabecalho = {...configPagina,"Authorization": "Bearer " + token};
+                            let opcoes =  {method: "DELETE",body: empresa,headers: cabecalho};
+                            let resposta = await fetch(urlEmpresa + "/" + empresa.id,opcoes);
+                            let msg = await obtemMensagemErro(resposta,props.navigation);
+                    
+                            if (msg && msg !== "")
+                                throw new Error(msg);
+
+                            await obtemEmpresas();
+                            
+                            msg = "Empresa excluída com sucesso.";
+                            Alert.alert("Empresa",msg);
+                        }
                     }
                     catch (erro) {
                         Alert.alert("Empresa",JSON.parse(erro.message).mensagem);
+                    }
+                    finally {
+                        setEsperar(false);
                     }
                 }
             },
@@ -124,6 +147,13 @@ export default function ListaEmpresas(props) {
       
     return (
         <Provider>
+            {
+                esperar
+                ?
+                    <Espera />
+                :
+                    undefined
+            }
             <SafeAreaView style={estilo.painel}>
                 <Menu visible={stateMenu.menuVisivel} onDismiss={() => menuInativo(dispatchMenu)} anchor={criaMenu(menuAtivo,dispatchMenu,props.navigation,"Empresa")}>
                     <Menu.Item onPress={() => alteraTela("ListaEmpresas")} title="Empresas" />
